+++ src/utils/vue/effectDispose.ts
@@ -19,7 +19,7 @@ import type { AnyFC } from '@/types/modules/utils'
  *
  * @remark 返回 true 表示获取到 effect 作用域并且卸载；false 表示未存在 effect 作用域
  */
-export function scopeDispose(fc: AnyFC) {
+export function effectDispose(fc: AnyFC) {
   if (getCurrentScope()) {
     onScopeDispose(fc)